final commit

+++ database/migrations/2024_11_02_181735_create_games_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->uuid() ;
             $table->string('title') ;
-            $table->double('size') ;
+            $table->float('size') ;
             $table->string('description') ;
             $table->json('image') ;
             $table->string('url_video') ;
